feat: add ignore rec button and ignore list functionality
