--- brbot/Data/anime_rec_cache.json
@@ -1,1 +1,0 @@
-{"5862798":{"date":"2025/05/28 00:21:53","recs":{"20954":100.0,"21234":73.48183523026589,"20722":72.7535495910514,"11061":70.4315348239337,"99750":69.3050272514266,"918":63.19456138522076,"128545":62.62494301878083,"20668":62.348729198539644,"6045":61.74612729479366,"9989":60.85492955291522,"20997":60.7559839421892,"141774":59.93131941285404,"21507":58.77845796047647,"101001":56.803489439857245,"12189":56.67027833950304,"153152":54.763390259130965,"11843":54.72383640256307,"20665":54.58634667335034,"30":53.583040023951625,"98460":52.485131008589335}},"5972899":{"date":"2025/05/28 00:26:16","recs":{"101347":100.0,"108268":94.84136289228042,"2608":93.4632141108381,"134732":91.27238516557053,"155907":90.47022404504462,"20464":86.87653212154481,"918":86.48686389875475,"437":86.18896965462041,"16498":85.88652680207589,"153518":84.08812846445566,"20755":83.71674216086954,"339":81.32269613757147,"11981":81.21138131827863,"21804":81.11661442021112,"2251":78.8613638307026,"19":78.04270662599116,"21450":77.89376939919828,"486":77.68927250090142,"4282":76.95308537976092,"20668":76.71967082149948}}}
