+++ brbot/Data/ignored_recs.json
@@ -1,0 +1,1 @@
+{"302266697488924672":{"manga":[],"anime":[]}}
