--- brbot/Data/manga_rec_cache.json
@@ -1,1 +1,0 @@
-{}
