--- brbot/Data/recommendation_cache.json
@@ -1,1 +1,0 @@
-{"5862798":{"date":"2025/05/21 22:34:19","recs":{"20954":100.0,"21234":83.18728894648112,"20722":82.22672635184603,"11061":81.4207048322062,"99750":80.98082312086333,"918":76.51254911548344,"128545":74.55718898656647,"141774":72.93361645100116,"20997":72.49341116569823,"20668":71.39755982845945,"9989":70.84425212836067,"21507":70.42185821906104,"6045":69.57821055452763,"20665":68.90648306048614,"11843":68.10357957575239,"12189":67.69204462535873,"101001":67.54083854909861,"30":67.33943876109757,"153152":66.29581194999618,"877":65.02636401631999}},"6451469":{"date":"2025/05/21 22:37:47","recs":{"14713":100.0,"21827":97.3234170891186,"9253":97.21515776244604,"127412":94.4874101456372,"853":93.99206323990047,"128715":92.25588354474688,"113415":87.69567172956599,"6045":87.39583055720969,"99578":87.29739225770541,"235":84.48198980080723,"123802":84.21718433536765,"16592":83.62014306349452,"105333":80.97836181817016,"99750":80.94928611057759,"143338":80.72016039670854,"21366":80.3947952300313,"101291":79.35564003078778,"122052":78.754723835562,"98436":76.3165717785668,"98478":76.00726348382781}},"940264":{"date":"2025/05/21 22:37:21","recs":{"9253":100.0,"21087":87.47435913737527,"14713":85.7443823944067,"97940":83.60822739013783,"20770":82.95515754218177,"1575":82.58232667628566,"12189":80.4014598787723,"101347":78.98489408602538,"142877":77.51664736521737,"130003":76.35585161697745,"11887":74.17240467170443,"168194":73.29674026181576,"118399":70.63565249126319,"232":69.89661246610486,"440":69.78658987111396,"392":69.62601628284558,"14741":68.87435536331398,"21804":68.34507872847992,"16498":68.28630625212475,"205":67.95071318219811}}}
